image_processor_lambda/requirements.txt
aws-lambda-powertools
ulid-py